added create device script

+++ driver_test.sh
@@ -8,6 +8,21 @@
 #mknod my_char_dev c 60 0
 #mknod my_char_dev_return c 61 0
 
+echo "compiling readInodes"
+g++ -std=c++11 -o readInodes readInodes.cc
+if [[ ! $? ]]; then
+	echo "readInodes compilation failed...exiting"
+	exit
+fi
+
+echo "compiling write nodes"
+g++ -std=c++11 -o writenodes writenodes.cc
+
+if [[ ! $? ]]; then
+	echo "writenodes compilation failed...exiting"
+	exit
+fi
+
 rm -rf my_char_dev my_char_dev_return &> /dev/null
 
 
@@ -23,8 +38,16 @@ cp my_char_dev.c my_char_dev/my_char_dev.c
 cd my_char_dev
 #rm *.mod.o *.mod.c *.o *.ko Module.symvers
 make
+if [[ ! $? ]]; then
+	echo "make for my_char_dev failed...exiting"
+	exit#statements
+fi
 echo "inserting my char dev now"
 insmod my_char_dev.ko
+if [[ ! $? ]]; then
+	echo " my_char_dev insertion failed ..exiting"
+	exit
+fi
 
 cd ..
 mkdir my_char_dev_return
@@ -34,8 +57,17 @@ cp my_char_dev_return.c my_char_dev_return/my_char_dev_return.c
 cd my_char_dev_return
 #rm *.mod.o *.mod.c *.o *.ko Module.symvers
 make
+if [[ ! $? ]]; then
+	echo "make for my_char_dev_return failed...exiting"
+	exit
+fi
+
 echo "inserting my_char_dev_return now"
 insmod my_char_dev_return.ko
+if [[ ! $? ]]; then
+	echo "my_char_dev_return insertion failed...exiting"
+	exit #statements
+fi
 
 cd ..
 
@@ -43,6 +75,10 @@ echo "Making read inode list now"
 
 cp Makefile.read_inode_list Makefile
 make
+if [[ ! $? ]]; then
+	echo " read_inode_list compilation failed...exiting"
+	exit
+fi
 #insmod read_inode_list.ko
 #rm -rf my_char_dev my_char_dev_return
